fixed difference between old greddy sample and generate
