+++ src/cpp/src/utils.cpp
@@ -49,5 +49,87 @@ std::pair<int64_t, float> softmax(const ov::Tensor& logits, const size_t batch_i
     return {out_token, log_sum};
 }
 
+void initialize_position_ids(ov::Tensor& position_ids, const ov::Tensor& attention_mask, int64_t start_pos) {
+    const size_t batch_size = attention_mask.get_shape()[0];
+    const size_t seq_length = attention_mask.get_shape()[1];
+
+    const int64_t* attention_mask_data = attention_mask.data<int64_t>();
+    int64_t* position_ids_data = position_ids.data<int64_t>();
+
+    for (size_t batch = 0; batch < batch_size; batch++) {
+        size_t sum = start_pos;
+        for (size_t i = 0; i < seq_length; i++) {
+            const size_t element_offset = batch * seq_length + i;
+            position_ids_data[element_offset] = sum;
+            if (attention_mask_data[element_offset] == 1) {
+                sum += 1;
+            }
+        }
+    }
+}
+
+void initialize_beam_inputs(const ov::Tensor& input_ids, const ov::Tensor& attention_mask, ov::InferRequest& request) {
+    request.set_tensor("input_ids", input_ids);
+    request.set_tensor("attention_mask", attention_mask);
+
+    ov::Shape input_shape = input_ids.get_shape();
+
+    ov::Tensor position_ids = request.get_tensor("position_ids");
+    position_ids.set_shape(input_shape);
+    initialize_position_ids(position_ids, attention_mask);
+
+    ov::Tensor beam_idx = request.get_tensor("beam_idx");
+    beam_idx.set_shape({input_shape.at(0)});
+    std::fill_n(beam_idx.data<int32_t>(), input_shape.at(0), 0);
+}
+
+
+void set_attention_mask(ov::Tensor&& attention_mask, std::vector<int32_t> next_beams) {
+    ov::Tensor original_mask{ov::element::i64, attention_mask.get_shape()};
+    ov::Shape original_shape = original_mask.get_shape();
+    attention_mask.copy_to(original_mask);
+
+    ov::Shape new_shape{next_beams.size(), original_mask.get_shape().at(1) + 1};
+    attention_mask.set_shape(new_shape);
+
+    for (size_t beam_id = 0; beam_id < next_beams.size(); beam_id++) {
+        const size_t original_prompt_offset = next_beams.at(beam_id) * original_shape.at(1);
+        const size_t result_prompt_offset = beam_id * new_shape.at(1);
+
+        int64_t* dest = attention_mask.data<int64_t>() + result_prompt_offset;
+        const int64_t* src = original_mask.data<int64_t>() + original_prompt_offset;
+
+        std::memcpy(dest, src, original_shape.at(1) * sizeof(int64_t));
+        attention_mask.data<int64_t>()[result_prompt_offset + new_shape.at(1) - 1] = 1;
+    }
+}
+
+void update_position_ids(ov::Tensor&& position_ids, const ov::Tensor&& attention_mask) {
+    const size_t batch_size = attention_mask.get_shape().at(0);
+    const size_t atten_length = attention_mask.get_shape().at(1);
+    position_ids.set_shape({batch_size, 1});
+
+    for (size_t batch = 0; batch < batch_size; batch++) {
+        int64_t* start = attention_mask.data<int64_t>() + batch * atten_length;
+        // todo: be careful with start + atten_length, probably need to replace with start + atten_length -1
+        position_ids.data<int64_t>()[batch] = std::accumulate(start, start + atten_length, 0);
+    }
+}
+
+ov::Tensor extend_attention(ov::Tensor attention_mask) {
+    auto shape = attention_mask.get_shape();
+    auto batch_size = shape[0];
+    auto seq_len = shape[1];
+
+    ov::Tensor new_atten_mask = ov::Tensor{attention_mask.get_element_type(), {batch_size, seq_len + 1}};
+    auto old_data = attention_mask.data<int64_t>();
+    auto new_data = new_atten_mask.data<int64_t>();
+    for (size_t batch = 0; batch < batch_size; ++batch) {
+        std::memcpy(new_data + batch * (seq_len + 1), old_data + batch * seq_len, seq_len * sizeof(int64_t));
+        new_data[batch * (seq_len + 1) + seq_len] = 1;
+    }
+    return new_atten_mask;
+}
+
 }  // namespace generate_utils
 }  // namespace ov
