+++ src/cpp/src/utils.hpp
@@ -14,6 +14,12 @@ void print_tensor(const ov::Tensor& tensor);
 
 std::pair<int64_t, float> softmax(const ov::Tensor& logits, const size_t batch_idx);
 
+void initialize_position_ids(ov::Tensor& position_ids, const ov::Tensor& attention_mask, int64_t start_pos = 0);
+
+ov::Tensor extend_attention(ov::Tensor attention_mask);
+
+void update_position_ids(ov::Tensor&& position_ids, const ov::Tensor&& attention_mask);
+
 bool is_xml(const std::string& path);
 
 }  // namespace generate_utils
