+++ text_generation/causal_lm/cpp/CMakeLists.txt
@@ -12,13 +12,10 @@ target_link_libraries(greedy_causal_lm PRIVATE generate_pipeline_lib)
 set_target_properties(greedy_causal_lm PROPERTIES CXX_STANDARD 17)
 set_target_properties(greedy_causal_lm PROPERTIES CXX_STANDARD_REQUIRED ON)
 
-# add_executable(beam_search_causal_lm beam_search_causal_lm.cpp)
-# target_compile_definitions(beam_search_causal_lm PRIVATE OPENVINO_TOKENIZERS_PATH=\"$<TARGET_FILE:openvino_tokenizers>\")
-# target_include_directories(beam_search_causal_lm PRIVATE ./)
-# find_package(OpenVINO REQUIRED COMPONENTS Runtime)
-# target_link_libraries(beam_search_causal_lm PRIVATE openvino::runtime)
-# set_target_properties(beam_search_causal_lm PROPERTIES CXX_STANDARD 17)
-# set_target_properties(beam_search_causal_lm PROPERTIES CXX_STANDARD_REQUIRED ON)
+add_executable(beam_search_causal_lm beam_search_causal_lm.cpp)
+target_link_libraries(beam_search_causal_lm PRIVATE generate_pipeline_lib)
+set_target_properties(beam_search_causal_lm PROPERTIES CXX_STANDARD 17)
+set_target_properties(beam_search_causal_lm PROPERTIES CXX_STANDARD_REQUIRED ON)
 
 add_executable(speculative_decoding_lm speculative_decoding_lm.cpp)
 target_compile_definitions(speculative_decoding_lm PRIVATE OPENVINO_TOKENIZERS_PATH=\"$<TARGET_FILE:openvino_tokenizers>\")
